chore(routesPage): API  versioning (#49)
-add an alias to routes
-version all the controllers
-fix all request tests

[Finish #161341636]
