+++ app/controllers/api/v1/application_controller.rb
@@ -1,0 +1,9 @@
+module Api
+  module V1
+    class ApplicationController < ErrorsController
+      include Pundit
+      include Knock::Authenticable
+    end
+  end
+end
+
