--- app/controllers/application_controller.rb
@@ -1,4 +1,0 @@
-class ApplicationController < ErrorsController
-  include Pundit
-  include Knock::Authenticable
-end
