+++ spec/requests/commentaries_spec.rb
@@ -49,7 +49,7 @@ RSpec.describe Commentary, type: :request do
   describe "POST commentary/create" do
     context "when the request is valid" do
       before do
-        post "/fixture/#{fixture_id}/commentaries",
+        post api_v1_add_commentary_path(fixture_id: fixture_id),
              headers: authenticated_header(user),
              params: commentary_params
       end
@@ -68,9 +68,9 @@ RSpec.describe Commentary, type: :request do
       let(:user) { create(:user, role_id: role.id) }
 
       before do
-        post "/fixture/#{fixture_id}/commentaries",
-             headers: authenticated_header(user),
-             params: commentary_params
+        post  api_v1_add_commentary_path(fixture_id: fixture_id),
+              headers: authenticated_header(user),
+              params: commentary_params
       end
 
       it "creates a new event" do
@@ -85,7 +85,9 @@ RSpec.describe Commentary, type: :request do
 
   describe "GET /fixture/:fixture_id/commentaries" do
     context "when the request is valid" do
-      before { get "/fixture/#{fixture_id}/commentaries" }
+      before do
+        get  api_v1_game_commentaries_path(fixture_id: fixture_id)
+      end
 
       it "returns a list with 10 hashes" do
         expect(json.size).to eq 10
@@ -99,7 +101,9 @@ RSpec.describe Commentary, type: :request do
     context "when the request is invalid" do
       let!(:fixture_id) { 1 }
 
-      before { get "/fixture/#{fixture_id}/commentaries" }
+      before do
+        get  api_v1_game_commentaries_path(fixture_id: fixture_id)
+      end
 
       it "returns an error message" do
         expect(json["error"]).to eq("No commentary for this game.")
@@ -117,7 +121,8 @@ RSpec.describe Commentary, type: :request do
       let(:user) { create(:user, role_id: role.id) }
 
       before do
-        delete "/fixture/#{fixture_id}/commentaries/#{commentary_id}",
+        delete api_v1_delete_commentary_path(fixture_id: fixture_id,
+                                             commentary_id: commentary_id),
                headers: authenticated_header(user)
       end
 
@@ -132,7 +137,8 @@ RSpec.describe Commentary, type: :request do
 
     context "when the request is made by an admin" do
       before do
-        delete "/fixture/#{fixture_id}/commentaries/#{commentary_id}",
+        delete api_v1_delete_commentary_path(fixture_id: fixture_id,
+                                             commentary_id: commentary_id),
                headers: authenticated_header(user)
       end
 
@@ -150,7 +156,8 @@ RSpec.describe Commentary, type: :request do
       let(:commentary_id) { 100 }
 
       before do
-        delete "/fixture/#{fixture_id}/commentaries/#{commentary_id}",
+        delete api_v1_delete_commentary_path(fixture_id: fixture_id,
+                                             commentary_id: commentary_id),
                headers: authenticated_header(user)
       end
 
@@ -167,7 +174,8 @@ RSpec.describe Commentary, type: :request do
   describe "PUT /commentaries/:commentary_id" do
     context "when the request is valid" do
       before do
-        put "/fixture/#{fixture_id}/commentaries/#{commentary_id}",
+        put api_v1_edit_commentary_path(fixture_id: fixture_id,
+                                        commentary_id: commentary_id),
             headers: authenticated_header(user)
       end
 
@@ -184,7 +192,8 @@ RSpec.describe Commentary, type: :request do
       let(:commentary_id) { 0 }
 
       before do
-        put "/fixture/#{fixture_id}/commentaries/#{commentary_id}",
+        put api_v1_edit_commentary_path(fixture_id: fixture_id,
+                                        commentary_id: commentary_id),
             headers: authenticated_header(user)
       end
