+++ spec/requests/fixtures_spec.rb
@@ -56,7 +56,7 @@ RSpec.describe Fixture, type: :request do
   describe "POST /league/:league_id/fixtures" do
     context "when the request is valid" do
       before do
-        post "/league/#{league_id}/fixtures",
+        post api_v1_add_fixtures_path(league_id: league_id),
              headers: authenticated_header(user),
              params: fixture_params
       end
@@ -74,7 +74,7 @@ RSpec.describe Fixture, type: :request do
   describe "GET /league/:league_id/fixtures" do
     context "when the request is valid" do
       before do
-        get "/league/#{league_id}/fixtures"
+        get api_v1_fixtures_path(league_id: league_id)
       end
 
       it "returns a list with 10 hashes" do
@@ -89,7 +89,8 @@ RSpec.describe Fixture, type: :request do
     context "when the request is invalid" do
       let!(:fixture_id) { 0 }
       before do
-        get "/league/#{league_id}/fixtures/#{fixture_id}"
+        get api_v1_show_fixture_path(league_id: league_id,
+                                fixture_id: fixture_id)
       end
 
       it "returns an error message" do
@@ -105,7 +106,8 @@ RSpec.describe Fixture, type: :request do
   describe "DELETE /league/:league_id/fixtures/:fixture_id" do
     context "when the request is valid" do
       before do
-        delete "/league/#{league_id}/fixtures/#{fixture_id}",
+        delete api_v1_delete_fixture_path(league_id: league_id,
+                                          fixture_id: fixture_id),
                headers: authenticated_header(user)
       end
 
@@ -122,7 +124,8 @@ RSpec.describe Fixture, type: :request do
       let(:fixture_id) { 0 }
 
       before do
-        delete "/league/#{league_id}/fixtures/#{fixture_id}",
+        delete api_v1_delete_fixture_path(league_id: league_id,
+                                          fixture_id: fixture_id),
                headers: authenticated_header(user)
       end
 
@@ -139,7 +142,8 @@ RSpec.describe Fixture, type: :request do
   describe "PUT /league/:league_id/fixtures/:fixture_id" do
     context "when the request is valid" do
       before do
-        put "/league/#{league_id}/fixtures/#{fixture_id}",
+        put api_v1_edit_fixture_path(league_id: league_id,
+                                       fixture_id: fixture_id),
             headers: authenticated_header(user)
       end
 
@@ -156,7 +160,8 @@ RSpec.describe Fixture, type: :request do
       let(:fixture_id) { 0 }
 
       before do
-        put "/league/#{league_id}/fixtures/#{fixture_id}",
+        put api_v1_edit_fixture_path(league_id: league_id,
+                                     fixture_id: fixture_id),
             headers: authenticated_header(user)
       end
