+++ spec/requests/sponsor_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe Sponsor, type: :request do
   describe "POST /sponsors" do
     context "when the request is valid" do
       before do
-        post "/sponsors",
+        post api_v1_add_sponsor_path,
              headers: authenticated_header(user),
              params: sponsor_params
       end
@@ -42,7 +42,7 @@ RSpec.describe Sponsor, type: :request do
       end
 
       before do
-        post "/sponsors",
+        post api_v1_add_sponsor_path,
              headers: authenticated_header(user),
              params: sponsor_params
       end
@@ -59,7 +59,9 @@ RSpec.describe Sponsor, type: :request do
 
   describe "GET /sponsors" do
     context "when the request is valid" do
-      before { get "/sponsors" }
+      before do
+        get api_v1_sponsors_path
+      end
 
       it "shows an array of 10 sponsors" do
         expect(json.size).to eq 10
@@ -73,7 +75,9 @@ RSpec.describe Sponsor, type: :request do
 
   describe "GET /sponsors/:sponsor_id" do
     context "when the request is valid" do
-      before { get "/sponsors/#{sponsor_id}" }
+      before do
+        get api_v1_sponsor_path(sponsor_id: sponsor_id)
+      end
 
       it "returns one sponsor's details" do
         expect(json.size).to eq 6
@@ -86,7 +90,9 @@ RSpec.describe Sponsor, type: :request do
 
     context "when the request is invalid" do
       let!(:sponsor_id) { 1000 }
-      before { get "/sponsors/#{sponsor_id}" }
+      before do
+        get api_v1_sponsor_path(sponsor_id: sponsor_id)
+      end
 
       it "returns an error message" do
         expect(json["errors"]).to eq("The sponsor does not exist")
@@ -102,7 +108,7 @@ RSpec.describe Sponsor, type: :request do
     context "when the request is valid" do
 
       before do
-        delete "/sponsors/#{sponsor_id}",
+        delete api_v1_delete_sponsor_path(sponsor_id: sponsor_id),
                headers: authenticated_header(user)
       end
 
@@ -119,7 +125,7 @@ RSpec.describe Sponsor, type: :request do
       let(:sponsor_id) { 100 }
 
       before do
-        delete "/sponsors/#{sponsor_id}",
+        delete api_v1_delete_sponsor_path(sponsor_id: sponsor_id),
                headers: authenticated_header(user)
       end
 
@@ -136,7 +142,7 @@ RSpec.describe Sponsor, type: :request do
   describe "PUT /sponsors/:sponsor_id" do
     context "when the request is valid" do
       before do
-        put "/sponsors/#{sponsor_id}",
+        put api_v1_edit_sponsor_path(sponsor_id: sponsor_id),
             headers: authenticated_header(user)
       end
 
@@ -153,7 +159,7 @@ RSpec.describe Sponsor, type: :request do
       let(:sponsor_id) { 100 }
 
       before do
-        put "/sponsors/#{sponsor_id}",
+        put api_v1_edit_sponsor_path(sponsor_id: sponsor_id),
             headers: authenticated_header(user)
       end
