+++ spec/requests/teams_spec.rb
@@ -17,10 +17,10 @@ RSpec.describe Team, type: :request do
 
   describe "POST teams/create" do
     context "when the request is valid" do
-      before { post "/teams", params: team_params }
-
       before do
-        post "/teams", headers: authenticated_header(user), params: team_params
+        post api_v1_add_team_path,
+             headers: authenticated_header(user),
+             params: team_params
       end
 
       it "creates a new team with 7 attributes" do
@@ -43,8 +43,9 @@ RSpec.describe Team, type: :request do
       end
 
       before do
-        post "/teams", headers: authenticated_header(user),
-                       params: team_params
+        post api_v1_add_team_path,
+             headers: authenticated_header(user),
+             params: team_params
       end
 
       it "does not create a new team with empty team name" do
@@ -59,7 +60,7 @@ RSpec.describe Team, type: :request do
 
   describe "GET /teams" do
     context "when the request is valid" do
-      before { get "/teams" }
+      before { get api_v1_teams_path }
 
       it "shows an array of 10 teams" do
         expect(json.size).to eq 10
@@ -74,7 +75,7 @@ RSpec.describe Team, type: :request do
   describe "DELETE /team/:team_id" do
     context "when the request is valid" do
       before do
-        delete "/teams/#{team_id}",
+        delete api_v1_delete_team_path(team_id: team_id),
                headers: authenticated_header(user)
       end
 
@@ -88,7 +89,7 @@ RSpec.describe Team, type: :request do
       let(:team_id) { 0 }
 
       before do
-        delete "/teams/#{team_id}",
+        delete api_v1_delete_team_path(team_id: team_id),
                headers: authenticated_header(user)
       end
 
@@ -101,7 +102,9 @@ RSpec.describe Team, type: :request do
 
   describe "GET /team/:id" do
     context "when the request is valid" do
-      before { get "/teams/#{team_id}" }
+      before do
+        get api_v1_team_path(team_id: team_id)
+      end
 
       it "returns a hash with 7 keys" do
         expect(json.size).to eq 7
@@ -114,7 +117,9 @@ RSpec.describe Team, type: :request do
 
     context "when the request is invalid" do
       let(:team_id) { 100 }
-      before { get "/teams/#{team_id}" }
+      before do
+        get api_v1_team_path(team_id: team_id)
+      end
 
       it "returns an error message and status code 400" do
         expect(response).to have_http_status(400)
@@ -125,10 +130,9 @@ RSpec.describe Team, type: :request do
 
   describe "PUT /teams/edit/:id" do
     context "when the request is valid" do
-      before { put "/teams/#{team_id}" }
 
       before do
-        put "/teams/#{team_id}",
+        put api_v1_edit_team_path(team_id: team_id),
             headers: authenticated_header(user)
       end
 
@@ -145,7 +149,7 @@ RSpec.describe Team, type: :request do
       let(:team_id) { 100 }
 
       before do
-        put "/teams/#{team_id}",
+        put api_v1_edit_team_path(team_id: team_id),
             headers: authenticated_header(user)
       end
